refactor(fmt): Rename termcolor mod to buffer
