+++ src/fmt/writer/buffer/extern_impl.rs
