+++ src/fmt/writer/buffer/shim_impl.rs
