+++ src/fmt/writer/mod.rs
@@ -1,15 +1,15 @@
 mod atty;
-mod termcolor;
+mod buffer;
 
 use self::atty::{is_stderr, is_stdout};
-use self::termcolor::BufferWriter;
+use self::buffer::BufferWriter;
 use std::{fmt, io, mem, sync::Mutex};
 
 pub(super) mod glob {
     pub use super::*;
 }
 
-pub(super) use self::termcolor::Buffer;
+pub(super) use self::buffer::Buffer;
 
 /// Log target, either `stdout`, `stderr` or a custom pipe.
 #[non_exhaustive]
